@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 
 import { AngularFirestore, AngularFirestoreCollection, AngularFirestoreDocument } from '@angular/fire/firestore';
-import { Observable } from "rxjs/Observable";
+import { Observable } from 'rxjs';
 import { User, Roles } from "../models/user";
 import { map } from 'rxjs/operators';
 
