@@ -136,7 +136,7 @@ export class AuthService {
       uid: user.uid,
       email: user.email,
       displayName: user.displayName,
-      photoURL: user.photoURL,
+      // photoURL: user.photoURL,
       emailVerified: user.emailVerified,
       roles: {
       
